+++ Ressources/lien.txt
@@ -1,8 +1,11 @@
-https://medium.com/@apoddar573/making-your-own-custom-environment-in-gym-c3b65ff8cdaa
+https://medium.com/@apoddar573/making-your-own-custom-environment-in-gym-c3b65ff8cdaa /done
 
-https://github.com/openai/gym/blob/master/docs/creating-environments.md
+https://github.com/openai/gym/blob/master/docs/creating-environments.md /done
 
 https://github.com/openai/gym/blob/master/gym/core.py
 
 https://www.lemagit.fr/conseil/Reseaux-de-neurones-RNN-et-CNN-comment-les-differencier
 
+
+
+After you have installed your package with pip install -e gym-foo, you can create an instance of the environment with gym.make('gym_foo:foo-v0')
