--- rlfho/envs/foo_env.py
@@ -1,17 +1,0 @@
-import gym
-from gym import error, spaces, utils
-from gym.utils import seeding
-
-class FooEnv(gym.Env):
-  metadata = {'render.modes': ['human']}
-
-  def __init__(self):
-    pass
-  def step(self, action):
-    pass
-  def reset(self):
-    pass
-  def render(self, mode='human'):
-    pass
-  def close(self):
-    pass
