+++ setup.py
@@ -2,5 +2,5 @@ from setuptools import setup
 
 setup(name='rlfho',
       version='0.0.1',
-      install_requires=['gym']  # And any other dependencies foo needs
+      install_requires=['gym']  # And any other dependencies rlfho needs
 )
